plot sparse now works with matplotlib
