+++ python/src/scipp/plot/plot_2d.py
@@ -4,7 +4,9 @@
 
 # Scipp imports
 from . import config
-from .tools import axis_label, parse_colorbar, render_plot, Slicer
+from .render import render_plot
+from .slicer import Slicer
+from .tools import axis_label, parse_colorbar
 
 # Other imports
 import numpy as np
@@ -25,14 +27,12 @@ def plot_2d(input_data, axes=None, contours=False, cb=None, filename=None,
     particular dimension.
     """
 
-    # Get the variable inside the dataset
-    name, var = next(iter(input_data))
-
+    var = input_data[name]
     if axes is None:
         axes = var.dims
 
     # Parse colorbar
-    cbar = parse_colorbar(config.cb, cb, plotly=True)
+    cbar = parse_colorbar(cb, plotly=True)
 
     # Make title
     title = axis_label(var=var, name=name, log=cbar["log"])
